# swagger.yaml
openapi: 3.0.3

info:
  title: API_ProgWeb
  description: API para cadastro, listagem, atualização e exclusão de publicações.
  version: 1.0.0

servers:
  - url: http://localhost:3000

paths:
  /posts:
    get:
      summary: Listar todas as publicações
      responses:
        '200':
          description: Publicações listadas com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Publicacao'
    post:
      summary: Cadastrar uma nova publicação
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicacaoEntrada'
      responses:
        '201':
          description: Publicação criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publicacao'
        '400':
          description: Dados inválidos na requisição

  /posts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
    get:
      summary: Buscar uma publicação por ID
      responses:
        '200':
          description: Publicação encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publicacao'
        '404':
          description: Publicação não encontrada
    put:
      summary: Atualizar uma publicação por ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PublicacaoEntrada'
      responses:
        '200':
          description: Publicação atualizada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Publicacao'
        '400':
          description: Dados inválidos na requisição
        '404':
          description: Publicação não encontrada
    delete:
      summary: Excluir uma publicação por ID
      responses:
        '204':
          description: Publicação excluída com sucesso
        '404':
          description: Publicação não encontrada

components:
  schemas:
    PublicacaoEntrada:
      type: object
      required:
        - titulo
        - descricao
        - autor
        - dataPublicacao
        - fotoAutor
      properties:
        titulo:
          type: string
          example: Lançamento do novo sistema
        descricao:
          type: string
          example: Confira as novidades da plataforma.
        autor:
          type: string
          example: Maria Silva
        dataPublicacao:
          type: string
          format: date
          example: '2026-04-28'
        fotoAutor:
          type: string
          format: uri
          example: https://placehold.co/120x120?text=MS
    Publicacao:
      allOf:
        - type: object
          properties:
            id:
              type: integer
              example: 1
        - $ref: '#/components/schemas/PublicacaoEntrada'